+++ opensilex-sparql/src/main/java/org/opensilex/sparql/service/SPARQLService.java
@@ -241,9 +241,9 @@ public class SPARQLService extends BaseService implements SPARQLConnection, Serv
     @Override
     public void executeUpdateQuery(UpdateBuilder update) throws SPARQLException {
         addPrefixes(update);
-//        if (LOGGER.isDebugEnabled()) {
-//            LOGGER.debug("SPARQL UPDATE\n" + update.build().toString());
-//        }
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("SPARQL UPDATE\n" + update.buildRequest().toString());
+        }
         connection.executeUpdateQuery(update);
     }